+++ cnd/src/swap_protocols/han.rs
@@ -1,20 +1,46 @@
-use crate::swap_protocols::{
-    rfc003::{
-        create_swap::{HtlcParams, SwapEvent},
-        events::{
-            Deployed, HtlcDeployed, HtlcFunded, HtlcRedeemed, HtlcRefunded, Redeemed, Refunded,
+use crate::{
+    asset,
+    btsieve::ethereum::{Cache, Web3Connector},
+    htlc_location, identity,
+    swap_protocols::{
+        han, ledger,
+        rfc003::{
+            create_swap::{HtlcParams, SwapEvent},
+            events::{
+                Deployed, HtlcDeployed, HtlcFunded, HtlcRedeemed, HtlcRefunded, Redeemed, Refunded,
+            },
+            LedgerState,
         },
-        LedgerState,
+        state, LedgerStates, LocalSwapId, Role,
     },
-    state, LocalSwapId,
+    transaction,
 };
-use chrono::NaiveDateTime;
+use chrono::{NaiveDateTime, Utc};
 use futures::future::{self, Either};
 use genawaiter::{
     sync::{Co, Gen},
     GeneratorState,
 };
 use std::sync::Arc;
+use tracing_futures::Instrument;
+
+pub async fn new_han_ethereum_ether_swap(
+    swap_id: LocalSwapId,
+    connector: Arc<Cache<Web3Connector>>,
+    ethereum_ledger_state: Arc<LedgerStates>,
+    htlc_params: HtlcParams<ledger::Ethereum, asset::Ether, identity::Ethereum>,
+    role: Role,
+) {
+    han::create_watcher::<_, _, _, _, htlc_location::Ethereum, _, transaction::Ethereum>(
+        connector.as_ref(),
+        ethereum_ledger_state,
+        swap_id,
+        htlc_params,
+        Utc::now().naive_local(),
+    )
+    .instrument(tracing::error_span!("alpha_ledger", swap_id = %swap_id, role = %role))
+    .await
+}
 
 /// Returns a future that tracks the swap negotiated from the given request and
 /// accept response on a ledger.
@@ -27,7 +53,7 @@ use std::sync::Arc;
 ///
 /// It is highly unlikely for Bob to fund the HTLC now, yet the current
 /// implementation is still waiting for that.
-pub async fn create_watcher<C, S, L, A, H, I, T>(
+async fn create_watcher<C, S, L, A, H, I, T>(
     ethereum_connector: &C,
     ledger_state: Arc<S>,
     swap_id: LocalSwapId,
